fix: back out some of previous class moves
