+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/FTBChunksClient.java
@@ -28,11 +28,7 @@ import dev.ftb.mods.ftbchunks.client.gui.LargeMapScreen;
 import dev.ftb.mods.ftbchunks.client.gui.WaypointEditorScreen;
 import dev.ftb.mods.ftbchunks.client.map.*;
 import dev.ftb.mods.ftbchunks.client.map.color.ColorUtils;
-import dev.ftb.mods.ftbchunks.client.mapicon.EntityIcons;
-import dev.ftb.mods.ftbchunks.client.mapicon.EntityMapIcon;
-import dev.ftb.mods.ftbchunks.client.mapicon.MapIconComparator;
-import dev.ftb.mods.ftbchunks.integration.InWorldMapIcon;
-import dev.ftb.mods.ftbchunks.integration.TrackedPlayerMapIcon;
+import dev.ftb.mods.ftbchunks.client.mapicon.*;
 import dev.ftb.mods.ftbchunks.net.PartialPackets;
 import dev.ftb.mods.ftbchunks.net.SendGeneralDataPacket.GeneralChunkData;
 import dev.ftb.mods.ftblibrary.config.StringConfig;
